backend/analytics/requirements.txt
psycopg2-binary==2.9.9
PyJWT==2.8.0
XlsxWriter==3.1.9
